updat and fix probleme for frontend test
